Backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
psycopg[binary]
python-dotenv==1.0.1